Initial refactoring for pip packaging and PyPi distribution
