+++ cyber_radar/playbooks/grab_ftp_file_list.py
@@ -13,23 +13,21 @@
 #
 #  You should have received a copy of the GNU General Public License
 #  along with RADAR.  If not, see <https://www.gnu.org/licenses/>.
-from ftplib import FTP, error_perm
+
+from ftplib import FTP
 
 
 def run(target: dict):
     target_host = target.get('target_host', None)
     if not target_host:
-        return f'!!!  Scan Anonymous FTP failed, no target_host specified: {target}'
-    try:
-        ftp_connection = FTP(target_host)
-        ftp_connection.login()  # Login anonymous
+        return f'!!!  FTP directory listing failed, no target_host specified: {target}'
+    ftp_connection = FTP(target_host)
+    ftp_connection.login()
+    files = ftp_connection.nlst()
 
-        target_vulns = target.get('vulnerabilities', None)
-        if target_vulns:
-            target_vulns.append('Anonymous FTP Login')
-        else:
-            target['vulnerabilities'] = ['Anonymous FTP Login']
+    # Add info to metadata
+    if not target.get('details', None):
+        target['details'] = {}
+    target['details']['ftp_server_contents'] = files
+    return f'$$$  These files/directories were on the anonymous FTP server: {files}'
 
-        return f'$$$  {target_host} is vulnerable to Anonymous FTP Login'
-    except (ConnectionRefusedError, error_perm):
-        return
